+++ NEWS.md
@@ -4,6 +4,7 @@
 - Prometheus metrics are available in Posit Workbench (rstudio-pro:#3273)
 - Update to Electron 25.2.0 (#13322)
 - Additional support for publishing new content types to Posit Cloud (rstudio-pro#4541)
+- Add support Quarto Manuscript projects, a new Quarto feature that will be released in Quarto 1.4.
 
 ### Fixed
 - Fixed issue where Electron menubar commands are not disabled when modals are displayed (#12972)